Move LoopbackRelay to loopback.py
Moving this class is needed to be able to get test_ssh in as part of
the Python 3 port, without also having to port test_recvline.py at the same
time.
